pyproject.toml
[tool.poetry]
name = "orka-api"
version = "0.1.0"
description = ""
authors = ["Joe Black"]

[tool.poetry.dependencies]
python = "^3.7"
graphene-sqlalchemy = {git = "https://github.com/graphql-python/graphene-sqlalchemy.git"}
pymysql = "^0.9.3"
sqlalchemy = "^1.3"
sqlalchemy_utils = "^0.33.11"
cognitojwt = "^1.1"
flask = "^1.0"
flask_graphql = "^2.0"
flask_sqlalchemy = "^2.4"
flask_cors = "^3.0"
requests = "^2.21"
slugify = "^0.0.1"
munch = "^2.3"
zappa = "^0.48.2"
marshmallow = "^2.19"
graphql-core = {path = "vendor/graphql-core"}
graphene = {path = "vendor/graphene"}

[tool.poetry.dev-dependencies]
pylint = "^2.3"
pytest = "^4.4"
snapshottest = "^0.5.0"
pyyaml = "^5.1"
oyaml = "^0.9.0"


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
